cleaning up references in test fence create moving create_client to fence create to help resolve circular dependencies defining generic functions in config moving send email to google monitor to help with circular dependencies adding utils test
