+++ fence/scripting/google_monitor.py
@@ -7,6 +7,7 @@ handle invalid service accounts and projects.
 """
 import traceback
 
+import requests
 from gen3cirrus.google_cloud.iam import GooglePolicyMember
 from gen3cirrus import GoogleCloudManager
 from gen3cirrus.google_cloud.errors import GoogleAPIError
@@ -37,7 +38,7 @@ from fence.resources.google.access_utils import (
 from fence import utils
 from fence.config import config
 from fence.models import User
-from fence.errors import Unauthorized
+from fence.errors import Unauthorized, NotFound
 
 logger = get_logger(__name__)
 
@@ -443,6 +444,52 @@ def _get_user_email_list_from_google_project_with_owner_role(project_id):
                 if role.name.upper() == "OWNER"
             }
         )
+
+
+def send_email(from_email, to_emails, subject, text, smtp_domain):
+    """
+    Send email to group of emails using mail gun api.
+
+    https://app.mailgun.com/
+
+    Args:
+        from_email(str): from email
+        to_emails(list): list of emails to receive the messages
+        text(str): the text message
+        smtp_domain(dict): smtp domain server
+
+            {
+                "smtp_hostname": "smtp.mailgun.org",
+                "default_login": "postmaster@mailgun.planx-pla.net",
+                "api_url": "https://api.mailgun.net/v3/mailgun.planx-pla.net",
+                "smtp_password": "password", # pragma: allowlist secret
+                "api_key": "api key" # pragma: allowlist secret
+            }
+
+    Returns:
+        Http response
+
+    Exceptions:
+        KeyError
+
+    """
+    if smtp_domain not in config["GUN_MAIL"] or not config["GUN_MAIL"].get(
+        smtp_domain
+    ).get("smtp_password"):
+        raise NotFound(
+            "SMTP Domain '{}' does not exist in configuration for GUN_MAIL or "
+            "smtp_password was not provided. "
+            "Cannot send email.".format(smtp_domain)
+        )
+
+    api_key = config["GUN_MAIL"][smtp_domain].get("api_key", "")
+    email_url = config["GUN_MAIL"][smtp_domain].get("api_url", "") + "/messages"
+
+    return requests.post(
+        email_url,
+        auth=("api", api_key),
+        data={"from": from_email, "to": to_emails, "subject": subject, "text": text},
+    )
 
 
 def _send_emails_informing_service_account_removal(
